{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Django = "<3.3"
bcrypt = "^3.2.2"
django-behaviors = "^0.5.1"
django-environ = "^0.9.0"
django-ipware = "^4.0.2"
django-axes = "^5.35.0"
whitenoise = "^6.2.0"
django-storages = "^1.12.3"
djangorestframework = "^3.13.1"
djangorestframework-camel-case = "^1.3.0"
drf-jwt = "^1.19.2"
drf-spectacular = {extras = ["sidecar"], version = "^0.22.1"}
django-filter = "^22.1"
django-split-settings = "^1.1.0"
django-healthchecks = "^1.5.0"
redis = "^4.3.4"
sentry-sdk = "^1.6.0"
Pillow = "^9.2.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"
pytest-django = ">=3.9"
pytest-deadfixtures = "^2.2.1"
pytest-env = "^0.6.2"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.8.2"
pytest-randomly = "^3.12.0"
dotenv-linter = "^0.3.0"
freezegun = "^1.2.1"
mixer = "^7.2.2"
jedi = "^0.18.1"
flake8-absolute-import = "^1.0.0"
flake8-aaa = "^0.12.2"
flake8-bugbear = "^22.7.1"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-django = "^1.1.2"
flake8-eradicate = "^1.2.1"
flake8-isort = ">=4.0.0"
flake8-fixme = "^1.1.1"
flake8-mock = "^0.3"
flake8-multiline-containers = "^0.0.19"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^1.3.0"
flake8-pie = "^0.15.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-pytest = "^1.3"
flake8-pytest-style = "^1.6.0"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.2"
flake8-todo = "^0.7"
flake8-use-fstring = "^1.3"
flake8-variables-names = "^0.0.5"
flake8-walrus = "^1.1.0"
mypy = "^0.961"
django-stubs = "^1.12.0"
djangorestframework-stubs = "^1.7.0"
types-freezegun = "^1.1.10"
types-Pillow = "^9.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
